First update related to Vim 7.4
- First update related to Vim 7.4.
- Changed license text.
- Fixed issue with 'hidden'.  If 'hidden' is set, make sure that
g:bufExplorerFindActive is set to 0.  Otherwise, when using /bs or /bv,
and selecting a buffer, the original buffer will be switched to instead
of being opened in the newly created windows.
- Added new 'b' mapping when the bufExplorer window is opened.  When 'b'
is pressed, the user is prompted for the buffer number to switch to, and
is is then switched to when <CR> is pressed.  This allows for somewhat
faster buffer switching instead of using the j and k keys or the mouse
to select the buffer to switch to.
- Removed 'set nolist' from the Initialize() function as well as the
restore of the 'list' setting in the CleanUp() function.  These were
causing issues when multiple new files were opened from the command
line.  Furthermore, there was really no reason, that I can remember, to
why the 'list' setting was saved, modified, and restored anyways.
- Fixed issue with WinManager integration code not working correctly
anymore.
- Brought back the xnoremap setup for the 'd' and 'D' keys.  These were
removed for some reason after version 7.2.8.
- Thanks to all the contributors and testers.