Change "noremap" to "nnoremap" to avoid unexpected mapping modes.
The "noremap" command puts mappings in place for normal, visual, select,
and operator-pending modes.  This is unnecessarily broad and can cause
problems, especially for select mode.  The select-mode behavior Vim
promises[1] (and therefore users expect) is that typing any printable
character should cause the selection to be deleted, insert mode to be
entered, and the typed character to be inserted into the buffer.  When a
select-mode mapping begins with a printable character, the expected
behavior is no longer provided.  This can surprise users by performing
an action instead of entering their text, or by delaying their
keystrokes until the mapping cannot be matched.

Instead of "noremap", the "nnoremap" command creates only a normal-mode
mapping, which is the expected behavior for a command-like mapping.

[1]: See ":help select-mode" for details.