Change behavior of Toggle when BufExplorer is open.
1. If BufExplorer is already open in another window, go to it and keep it open.

2. If BufExplorer is already open in the current window, don't call BufExplorer() first, because doing so displays [BufExplorer] in the buffer list when it is opened again.

The root cause of the error appears to be the BufExplorer() function's use of the drop statement. This fix doesn't completely solve the problem, but it does make it much less likely to happen. Now only scenario 1 above will cause the normally-unlisted [BufExplorer] buffer to be displayed in the list.