v7.4.19
- Merged all changes from github back into this version and tried to
sync
to the correct version number.
7.4.18 - Github
7.4.17 - Github.
7.4.16   August 14, 2017
- Thanks to Yubo Su for the patch that adds 'f, F, V, and v' commands.
With this change, the original 'f' command was remapped to 'a'.

The set of commands that can be used during exploring are:
F - Open selected buffer in another window above the current.
f - Open selected buffer in another window below the current.
V - Open the selected buffer in another window on the left of the
current.
v - Open the selected buffer in another window on the right of the
current.