Fix MRUList sort order after loading vim session
Loading a session which includes hidden or unloaded buffers breaks the
MRUList sort order in BufExplorer, because only the visible buffers are
added to the MRUList and the MRUCmp() function doesn't deal with buffer
numbers not found in the MRUList, which causes visible buffers included
in the MRUList to be listed after other buffers not yet in the MRUList.

Fix this by simply resetting the MRUList after loading a session so all
listed buffers are included.

To allow for wiping out buffers when switching between sessions, which
is done by sessionman.vim and possibly other plugins and scripts, also
filter MRUList when resetting to exclude buffers that no longer exist.