Improve the display of terminal buffers.
Avoid calculating absolute and relative paths from the irregular path
syntax used for terminal buffers; instead, synthesize a path based on
the terminal's current working directory and the name of the shell:

  /current/working/directory/!PID:shellName

Where `/current/working/directory/` is the shell's current working
directory (when it can be determined), `PID` is the shell's process
identifier, and `shellName` is the name of the shell with leading path
removed for brevity.  For example, a terminal running `/bin/bash` as PID
1234 in directory `/tmp` would have a path of:

  /tmp/!1234:bash

On Neovim, terminal buffers are named as:

  term://WorkingDirectoryAtTerminalLaunch//PID:shellPath

On Vim, the name contains only a `!` and the shell path:

  !/path/to/shell

However, the PID is available on Vim via `term_getjob()` and `job_info()`.

At present, the current working directory of the shell can be determined
only on Unix-like systems having a `/proc` filesystem.  For other
systems, fall back to the directory at terminal launch for Neovim, or
else on Vim's current working directory.

It appears to be difficult to determine the current working directory of
a program on Windows given the PID, so no attempt is made to determine
it for that platform.