Remove visual-mode mappings for BufExplorer window.
These mappings for `d` and `D` may have been intended to support
selecting multiple buffers at once and deleting them as a group; at
present, however, support for this is incomplete.  In addition, it's
difficult to enter visual mode in the normal manner, as both `v` and `V`
have been redefined for other purposes.  CTRL-v is still available for
starting visual block mode, but this is less commonly known.

Support for the use case of "select multiple and delete as a group" has
been requested in issue #4, "Delete multiple buffers removes the wrong
ones"; if that feature is to be provided, it will likely be worked in
the context of that issue.