Use `X` to toggle the showing of terminal buffers.
When the feature to toggle showing terminal buffers was introduced, a
mapping was added for `H` to perform this toggling in the BufExplorer
window; however, the `H` key was immediately mapped again by
BufExplorer's logic to use `:keepjumps` with some movement keys (`H`,
`M`, `L`, etc.), thus disabling the intended `H` mapping:

  ```vim
  nnoremap <script> <silent> <nowait> <buffer> H             :call <SID>ToggleShowTerminal()<CR>

  for k in ["G", "n", "N", "L", "M", "H"]
      execute "nnoremap <buffer> <silent>" k ":keepjumps normal!" k."<CR>"
  endfor
  ```

As these movement keys are useful for navigating the BufExplorer window,
use `X` instead of `H` to toggle the showing of terminal buffers, as `X`
is normally an editing feature (for deleting characters before the
cursor) that's not needed in the BufExplorer window.

In addition, the `X` command has been documented in `bufexplorer.txt`
and in BufExplorer's `<F1>` help text.