Allow deletion/wiping of Neovim terminal buffers.
For terminal buffers, both `:bd` and `:bw` require `!` to force removal.
On Vim, terminal buffers appear to be modified (so `&modified == 1`);
however, Neovim terminal buffers have `&modified == 0`, so BufExplorer
was not previously asking the user for confirmation nor using `!` during
buffer removal, causing removal to fail with errors such as:

```
Vim(bdelete):E89: term://~/testing//816579:/bin/bash will be killed (add ! to override)
```

To allow removal, require confirmation for terminal buffers and adjust
the confirmation prompt for clarity.  This addresses the majority of
issue #69, "Feature Request: Add an option to do :bd!", as it's now
possible to remove terminal buffers; however, the issue hints at a
desire to configure BufExplorer to force-remove buffers unconditionally,
which is not (yet) possible.