Retain buffer until it has been fully wiped.
When deleting a buffer with `:bd`, the buffer becomes unlisted.  This
means the buffer should remain in the raw buffer list (though with
updated attributes), and if unlisted buffers are being displayed, the
buffer should stay visible.  It's also possible that wiping a buffer
might fail for some reason, in which case the buffer should be retained
in the buffer list.

After attempting to remove a buffer (either via `:bd` or `:bw`), if the
buffer still exists, regather information on that buffer and update the
display accordingly.