Improve error message when deleting an unlisted buffer.
Deleting a listed buffer makes it unlisted.  Deleting an unlisted buffer
would do nothing, so `:bd` results in an error message from Vim such as:

```
Vim(bdelete):E516: No buffers were deleted: silent bdelete 1
```

Improve BufExplorer's error message in this case, letting the user know
that this is an unlisted buffer that must be wiped to remove it.