Add `O` command to open in the original window at BufExplorer launch.
When launching BufExplorer via `:BufExplorer`, the original launch
window is reused temporarily for BufExplorer; then, when pressing
`<Enter>` to select a buffer, BufExplorer closes and the original launch
window is used for the selected buffer.

But when launching BufExplorer via `:BufExplorerHorizontalSplit` or
`:BufExplorerVerticalSplit`, a new split is created for BufExplorer.  If
`<Enter>` is used to select a buffer, this buffer will be opened in the
split created for BufExplorer.

This new `O` command provides a way to first close the split that was
created for BufExplorer at launch such that the selected buffer will
be opened in the original window.

This addresses issue #55.