Ignore user selection of a non-buffer line.
If the user causes `s:SelectBuffer()` to be called on a non-buffer line,
ignore their selection entirely, rather than moving the cursor down a
line.  Because there are multiple ways the user may trigger this
function call (e.g., double-clicking the mouse), moving the cursor can
have a jarring effect (e.g., making the user think the mouse jumped down
a line).  Performing no action helps reinforce to the user that there is
no buffer to select on this line.