Return to original window if possible.
When quitting BufExplorer via `q` after launching in a split, the
correct window is not always chosen.  With the default BufExplorer
options `g:bufExplorerSplitBelow = 0` and `g:bufExplorerSplitRight = 0`,
BufExplorer opens "before" (above or left-of) the original window, and
when BufExplorer closes via `q`, the correct (original) window is
chosen; however, with `g:bufExplorerSplitBelow = 1` and
`g:bufExplorerSplitRight = 1`, the wrong window is chosen (namely, the
window below or right-of the original).  To correct for this, the
window-at-launch is now activated when BufExplorer quits.